refactor: clean includes, move constructor to source file
